fix cypress test seeding
