+++ packages/cypress/package.json
@@ -1,7 +1,6 @@
 {
   "name": "oa-cypress",
   "version": "1.0.0",
-  "type": "module",
   "scripts": {
     "start": "cross-env PORT=3456 node --loader ts-node/esm scripts/start.mts"
   },
@@ -9,6 +8,7 @@
     "@faker-js/faker": "^7.6.0",
     "@types/fs-extra": "^11.0.4",
     "@types/wait-on": "^5.3.4",
+    "@vitejs/plugin-react": "^5.1.1",
     "cross-env": "^7.0.3",
     "cypress": "13.15.1",
     "cypress-file-upload": "5.0.8",
@@ -18,6 +18,9 @@
     "oa-shared": "workspace:*",
     "ts-node": "^10.9.2",
     "typescript": "^5.7.2",
+    "vite": "^7.2.2",
+    "vite-plugin-svgr": "^4.5.0",
+    "vite-tsconfig-paths": "^5.1.4",
     "wait-on": "^7.2.0",
     "wait-port": "^1.1.0"
   },
