+++ packages/cypress/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2020",
-    "module": "es2020",
+    "module": "commonjs",
     "lib": ["ESNext", "dom"],
     "types": ["cypress", "node", "cypress-file-upload"],
     "moduleResolution": "Node",
@@ -9,5 +9,11 @@
     "resolveJsonModule": true,
     "jsx": "react"
   },
-  "include": ["**/*.ts", "scripts/start.mts"]
+  "include": ["**/*.ts", "scripts/start.mts"],
+  "ts-node": {
+    "transpileOnly": true,
+    "compilerOptions": {
+      "module": "commonjs"
+    }
+  }
 }
